getting more pog

+++ src/displayapp/screens/Stopwatch.h
@@ -26,7 +26,7 @@ namespace Pinetime {
                 std::chrono::_V2::system_clock::time_point startTime;
                 std::chrono::_V2::system_clock::time_point endTime;
 
-                float elapsedTime = 0.;
+                double elapsedTime = 0.;
 
                 unsigned short int miliseconds;
                 unsigned short int seconds;
@@ -36,8 +36,8 @@ namespace Pinetime {
                 void startTimer();
                 void stopTimer();
                 void restartTimer();
-                void calculateTime(float timeDifference, char* timeStr);
-                void convertToHMS(float seconds, unsigned short int *ms, unsigned short int *s, unsigned short int *m, unsigned int *h);
+                void calculateTime(double timeDifference, char* timeStr);
+                void convertToHMS(double seconds, unsigned short int *ms, unsigned short int *s, unsigned short int *m, unsigned int *h);
 
                 lv_obj_t * label_time;
                 lv_obj_t * label_extra;
